Move context processing code to temporary `context2.js`.
